+++ app/controllers/meets_controller.rb
@@ -31,6 +31,58 @@ class MeetsController < ApplicationController
       end
     end
 
+    def available_meets
+      meets = Meet.includes(:availability_tutor, :users)
+                  .joins(:availability_tutor)
+
+      if params[:id_availability_tutor].present?
+        meets = meets.where(availability_tutor_id: params[:id_availability_tutor])
+      end
+
+      if params[:topic_id].present?
+        meets = meets.joins(availability_tutor: :topic)
+                     .where(availability_tutors: { topic_id: params[:topic_id] })
+      end
+
+      if params[:subject_id].present?
+        meets = meets.joins(availability_tutor: { topic: :subject })
+                     .where(topics: { subject_id: params[:subject_id] })
+      end
+
+      if params[:interested].present?
+        interested = params[:interested].to_s.downcase == "true"
+        if interested
+          meets = meets.joins(:participants).where(participants: { user_id: @current_user.id })
+        else
+          meets = meets.where.not(id: meets.joins(:participants).where(participants: { user_id: @current_user.id }))
+        end
+      end
+
+      if params[:meet_state].present?
+        meets = meets.where(status: map_meeting_status(params[:meet_state]))
+      end
+
+      meets_data = meets.map do |meet|
+        subject = meet.availability_tutor.topic.subject
+
+        {
+          id: meet.id,
+          topic_name: meet.availability_tutor.topic.name,
+          tutor_name: meet.availability_tutor.user.name,
+          meeting_date: meet.date_time,
+          meet_status: meet.status,
+          subject: {
+            id: subject.id,
+            name: subject.name
+          },
+          interested: meet.users.include?(@current_user),
+          number_of_interested: meet.users.size
+        }
+      end
+
+      render json: meets_data, status: :ok
+    end
+
     private
 
     # Find the AvailabilityTutor based on the ID in the URL
@@ -40,7 +92,6 @@ class MeetsController < ApplicationController
       render json: { error: "Availability tutor not found" }, status: :not_found
     end
 
-    # Find the Meet based on the ID in the URL
     def set_meet
       @meet = Meet.find(params[:id])
     rescue ActiveRecord::RecordNotFound
@@ -50,5 +101,20 @@ class MeetsController < ApplicationController
     # Find the date in timestamp in the URL
     def set_date
       @dateTS = params.require(:meet).permit[:date]
+    end
+
+    def map_meeting_status(status_param)
+      case status_param.to_i
+      when 0
+          "pending"
+      when 1
+          "confirmed"
+      when 2
+          "completed"
+      when 3
+          "canceled"
+      else
+          nil
+      end
     end
 end