+++ app/controllers/topics_controller.rb
@@ -13,6 +13,11 @@ class TopicsController < ApplicationController
       topics_by_subject = Topic.where(subject_id: params[:subject_id]).pluck(:id)
       availability_tutors = availability_tutors.where(topic_id: topics_by_subject)
     end
+    availability_tutors = availability_tutors.
+                          left_joins(:interesteds).
+                          select("availability_tutors.*, COUNT(interesteds.id) AS interested_count").
+                          group("availability_tutors.id").
+                          order(Arel.sql("MAX(CASE WHEN interesteds.user_id = '#{ActiveRecord::Base.connection.quote_string(@current_user.first.id)}' THEN 1 ELSE 0 END) DESC, availability_tutors.created_at DESC"))
 
     render json: format_topic_response(availability_tutors), status: :ok
   end
@@ -88,7 +93,6 @@ class TopicsController < ApplicationController
         topic = availability_tutor.topic
         user = availability_tutor.user
         subject = topic.subject
-
         {
           availability_id: availability_tutor.id,
           topic_name: topic.name,
@@ -96,7 +100,8 @@ class TopicsController < ApplicationController
           topic_description: topic.description,
           availability: availability_tutor.availability,
           availability_description: availability_tutor.description,
-          interesteds: availability_tutor.interesteds.count,
+          interesteds: availability_tutor.interested_count,
+          interested: availability_tutor.interesteds.exists?(user_id: @current_user.first.id),
           subject: {
             id: subject.id,
             name: subject.name